% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eal_9_annotations.R
\name{AnnotModUI}
\alias{AnnotModUI}
\alias{AnnotMod}
\title{AnnotModUI}
\usage{
AnnotModUI(id, div_id, site_id, rmv_id, default = NULL)

AnnotMod(
  input,
  output,
  session,
  globals,
  rv,
  rmv_id,
  site_id,
  ref,
  role = NULL,
  saved = NULL
)
}
\description{
module to document EML annotation
}
\section{Functions}{
\itemize{
\item \code{AnnotMod}: 
}}

