% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-app-options.R
\name{appOptions}
\alias{appOptions}
\title{appOptions}
\usage{
appOptions(input, output, session, globals, server)
}
\description{
server part of the appOptions module. Allow the user to change several settings in the app.
}
