% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-app-options.R
\name{appOptionsUI}
\alias{appOptionsUI}
\title{appOptionsUI}
\usage{
appOptionsUI(id, settingsVar, dev)
}
\description{
UI part of the appOptions module. Allow the user to change several settings in the app.
}
