% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_additional_HTML.R
\name{centered}
\alias{centered}
\title{centered}
\usage{
centered(...)
}
\arguments{
\item{...}{any HTML tag.}
}
\value{
A div with \code{style = "text-align: center"}
}
\description{
Returns the content styled as centered text.
}
