% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-additional_HTML.R
\name{checkFeedback}
\alias{checkFeedback}
\title{checkFeedback

relies on {shinyFeedback} to automate feedback on one input.}
\usage{
checkFeedback(
  input,
  id,
  condition = NULL,
  silent = FALSE,
  type = c("danger", "warning"),
  text = NULL
)
}
\description{
checkFeedback

relies on {shinyFeedback} to automate feedback on one input.
}
