% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-fill-fill_functions.R
\name{checkTruth}
\alias{checkTruth}
\title{checkTruth}
\usage{
checkTruth(x, output = NULL)
}
\arguments{
\item{x}{argument to check fo truthiness}

\item{output}{what to return if `x` is not truthy}
}
\description{
check if `x` is truthy (as shiny::isTruthy) or not. 
Returns the argument if truthy, or the `output` argument if not (default to NULL)
}
