% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-doc-documentation_functions.R
\name{commonPath}
\alias{commonPath}
\title{commonPath}
\usage{
commonPath(li, name)
}
\arguments{
\item{li}{lists in which path is searched}

\item{name}{currently selected node's name}
}
\description{
Extract the path made common from every leaf path. Used to get a
 non-leaf node's path
}
