% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eal_3_Attributes_dates.R
\name{convertLubridateFormat}
\alias{convertLubridateFormat}
\title{Convert lubridate to common format}
\usage{
convertLubridateFormat(date_formats)
}
\description{
Convert lubridate to common format
}
