% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL_functions.R
\name{createDPFolder}
\alias{createDPFolder}
\title{createDPFolder}
\usage{
createDPFolder(DP.location, DP.name)
}
\arguments{
\item{DP.location}{the directory path where to create the folder}

\item{DP.name}{the name to set on the folder}
}
\description{
if a folder is about to be created with same name and
location an existing folder owns, the existing folder is deleted
and a message appear through a shiny modal.
_(the 'DP' prefixing refers to 'data package')_
_(this function is not used in MetaShARK)_
}
