% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL_functions.R
\name{DataFilesFolder}
\alias{DataFilesFolder}
\title{create DP directory}
\usage{
DataFilesFolder(DP.location, DP.name, data.location)
}
\description{
EML assembly line convenience function for templating directories and
avoiding duplicating it. A similarly named and located directory will be whatever deleted.
}
