% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-upload-functions.R
\name{describeWorkflowUI}
\alias{describeWorkflowUI}
\alias{describeWorkflow}
\title{describeWorkflowUI}
\usage{
describeWorkflowUI(id, sources, targets)

describeWorkflow(input, output, session)
}
\description{
UI part of the script-data workflow-describing linking module (in upload).

server part of the script-data workflow-describing linking module (in upload).
}
\section{Functions}{
\itemize{
\item \code{describeWorkflow}: 
}}

