% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-doc-documentation.R
\name{documentation}
\alias{documentation}
\title{documentation}
\usage{
documentation(input, output, session)
}
\description{
This is the function in charge of the server part of 'documentation'
module
}
