% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL_DPfiles.R
\name{DPfiles}
\alias{DPfiles}
\title{DPfiles}
\usage{
DPfiles(input, output, session, savevar, globals)
}
\arguments{
\item{savevar}{global reactiveValue containing the saved information
entered by the user.}

\item{globals}{global list containing fixed setting values for the
app.}
}
\description{
server part of the EML AL module (step 2: select data files for
the data package)
}
