% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL_DPfiles.R
\name{DPfilesUI}
\alias{DPfilesUI}
\title{DPfilesUI}
\usage{
DPfilesUI(id, title, dev = FALSE)
}
\arguments{
\item{title}{page title}

\item{dev}{logical. Shall the dev items appear?}
}
\description{
UI part of the EML AL module (step 2: select data files for
the data package)
}
