% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL.R
\name{EMLAL}
\alias{EMLAL}
\title{EMLAL}
\usage{
EMLAL(input, output, session, savevar, globals)
}
\arguments{
\item{savevar}{global reactiveValue containing the saved information
entered by the user.}

\item{globals}{global list containing fixed setting values for the
app.}
}
\description{
server part of the 'EML Assembly Line' module. It relies
on work lead through the EDI tool 'EML Assembly Line'
}
