% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL.R
\name{EMLALUI}
\alias{EMLALUI}
\title{EMLALUI}
\usage{
EMLALUI(id, dev = FALSE)
}
\arguments{
\item{dev}{logical. Shall the dev items appear?}
}
\description{
UI part of the 'EML Assembly Line' module. It relies
on work lead through the EDI tool 'EML Assembly Line'
}
