% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-doc-documentation_functions.R
\name{extractContent}
\alias{extractContent}
\title{documentation_functions.R}
\usage{
extractContent(content, nsIndex)
}
\arguments{
\item{content}{node of the SystemGuideline matching the clicked node}

\item{nsIndex}{index of the used namespaces in the current module.}
}
\description{
extracts the targetted EML content from a node. This node is
selected in the documentation module
}
