% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL_GeoCovInput.R
\name{extractCoordinates}
\alias{extractCoordinates}
\title{extractCoordinates}
\usage{
extractCoordinates(coordCols, Pattern, localWarnings, coordTags, rv)
}
\description{
extractCoordinates
}
