% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-fill-fill.R
\name{fill}
\alias{fill}
\title{fill}
\usage{
fill(input, output, session, globals)
}
\arguments{
\item{globals}{global list containing fixed setting values for the
app.}
}
\description{
server part of the fill module
}
