% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-module.R
\name{fillUI}
\alias{fillUI}
\alias{fill}
\title{Fill}
\usage{
fillUI(id, dev = FALSE)

fill(input, output, session, globals)
}
\description{
server part of the fill module. Root script for whole module (comprise EML AL and MetaFIN)
}
\section{Functions}{
\itemize{
\item \code{fill}: 
}}

