% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL_geocovInput.R
\name{geoCovInputUI}
\alias{geoCovInputUI}
\alias{geoCovInput}
\title{Geographic coverage modular Input}
\usage{
geoCovInputUI(id, site_id, rmv_id, old_id)

geoCovInput(input, output, session, rv, rmv_id, site_id, ref)
}
\description{
UI part for the Geographic Coverage modular input, used
in Geographic Coverage module to input a custom entry in the dataset.
}
\section{Functions}{
\itemize{
\item \code{geoCovInput}: 
}}

