% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_grep.R
\name{\%grep\%}
\alias{\%grep\%}
\title{shorthand for base::grepl}
\usage{
x \%grep\% table
}
\arguments{
\item{x}{character. Pattern(s) to be matched against \code{table}.}

\item{table}{any unlisted data. A collection of data among which to find \code{pattern}-matching
elements.}
}
\description{
This function is designed to be a shorthand for base::grepl, thought to be used as \code{\%in\%}.
}
