% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-doc-documentation_functions.R
\name{has_child}
\alias{has_child}
\title{has_child

 @description check if one of the children of the input list is a list

 @param li list whose child(ren) are checked}
\usage{
has_child(li)
}
\description{
has_child

 @description check if one of the children of the input list is a list

 @param li list whose child(ren) are checked
}
