% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eal_9_annotations.R
\name{insertAnnotInput}
\alias{insertAnnotInput}
\title{insertPersonnelInput}
\usage{
insertAnnotInput(id, rv, ns, main.env, default = NULL)
}
\description{
helper function to insert PersonnelMod* functions. Calling this from
a shiny server will insert PersonnelModUI and create its server part. Provided with
features to delete them.
}
