% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eal_5_GeoCovInput.R
\name{insertGeoCovInput}
\alias{insertGeoCovInput}
\title{insertGeoCovInput}
\usage{
insertGeoCovInput(id, rv, ns)
}
\description{
Leads all the process of rendering a GeoCovInput
GUI and its associated server
}
