% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL_personnel.R
\name{insertPersonnelInput}
\alias{insertPersonnelInput}
\title{insertPersonnelInput}
\usage{
insertPersonnelInput(id, rv, ns, globals, role = NULL, saved = NULL)
}
\description{
helper function to insert PersonnelMod* functions. Calling this from
a shiny server will insert PersonnelModUI and create its server part. Provided with
features to delete them.
}
