% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-isContentTruthy.R
\name{isContentTruthy}
\alias{isContentTruthy}
\title{isContentTruthy}
\usage{
isContentTruthy(x)
}
\arguments{
\item{x}{argument to check fo truthiness}
}
\description{
check if `x` is truthy (as shiny::isTruthy) or not.
Returns the argument if truthy, or the `output` argument if not (default to NULL)
}
