% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_isContentTruthy.R
\name{isHTMLTruthy}
\alias{isHTMLTruthy}
\title{Check for HTML tags validity}
\usage{
isHTMLTruthy(x)
}
\arguments{
\item{x}{HTML tag. Target of the evaluation.}
}
\value{
A logical value indicating whether or not the tag is truthy.
}
\description{
Check if a HTML tag is well formed according to opening and closing chevrons.
}
