% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL_makeeml.R
\name{MakeEMLUI}
\alias{MakeEMLUI}
\alias{MakeEML}
\title{MakeEMLUI}
\usage{
MakeEMLUI(id, title, dev)

MakeEML(input, output, session, savevar, globals)
}
\description{
UI part for the Make EML module

server part of the make EML module
}
\section{Functions}{
\itemize{
\item \code{MakeEML}: 
}}

