% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-reactiveTrigger.R
\name{makeReactiveTrigger}
\alias{makeReactiveTrigger}
\title{makeReactiveTrigger}
\usage{
makeReactiveTrigger()
}
\description{
produces a trigger generator object X in 2 parts:
- X$depend() : written in a shiny server code chunk, this chunk comes to
be triggered each time X$trigger() is evaluated
- X$trigger() : written in a shiny server code chunk, each time X$trigger()
is evaluated, code chunks containing X$depend() will be evaluated after.
All creadits for this function goes to Dean Attali
}
