% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-markdownInput.R
\name{markdownInputUI}
\alias{markdownInputUI}
\alias{markdownInput}
\title{markdownInputUI}
\usage{
markdownInputUI(id, preview = FALSE, value = "")

markdownInput(input, output, session, preview = FALSE)
}
\arguments{
\item{id}{character. Shiny module id}

\item{preview}{logical. Shall a preview panel appear? (right-sided, 50\% width)}

\item{value}{character. Initial value of the text input (default to "")}
}
\value{
A HTML `reactive` containing the formatted input. Do call it as a reactive.
}
\description{
A shiny input module designed to allow the user to write markdown and render it as 
a HTML fix. 

It relies on the _Ace technology_ [https://ace.c9.io/#nav=about] with the package {shinyAce}.
}
\section{Functions}{
\itemize{
\item \code{markdownInput}: Server side of the module.
}}

