% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eal_8_misc.R
\name{MiscUI}
\alias{MiscUI}
\alias{Misc}
\title{MiscUI}
\usage{
MiscUI(id, title, dev, savevar)

Misc(input, output, session, savevar, globals, NSB)
}
\description{
UI for "last but not least" module
}
\section{Functions}{
\itemize{
\item \code{Misc}: 
}}

