% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-multiApply.R
\name{multiApply}
\alias{multiApply}
\title{multiApply}
\usage{
multiApply(target, FUNS = list(), ARGS = rep(NA, length(FUNS)),
  path = "Root", setPath = TRUE)
}
\arguments{
\item{target:}{the list on which the FUNS will be applied}

\item{FUNS:}{list of functions}

\item{ARGS:}{list of arguments matching FUNS. Caution ! The elements in ARGS
shall be concatenated in a vector if possible. Else, send a
pointer (list names or so). Remember you do not need to input
target in the ARGS list}

\item{path:}{if a FUN explicitly needs the path followed from the 0th-level,
it can be accessed by figuring it clearly as 'path' in the function
arguments.}
}
\description{
Can be used as template for recursive list function
}
