% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-multiFilesInput.R
\name{multiFIlesInput}
\alias{multiFIlesInput}
\title{multiFilesInput}
\usage{
multiFIlesInput(input, output, session)
}
\arguments{
\item{input}{shiny module input}

\item{output}{shiny module output}

\item{session}{shiny module session}
}
\description{
server part for the multiFilesInput module
}
