% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-multiFilesInput.R
\name{multiFIlesInputUI}
\alias{multiFIlesInputUI}
\title{multiFIlesInputUI}
\usage{
multiFIlesInputUI(id, helpText = NULL)
}
\arguments{
\item{id}{shiny module id}

\item{helpText}{a character vector to give the user an explanation. Will be inserted into a tags$p() call.}
}
\description{
UI part of the multiFilesInput. Allow the user to chose multiple files, and remove all or part of his selection.
}
