% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL_functions.R
\name{navSidebar}
\alias{navSidebar}
\alias{quitButton}
\alias{saveButton}
\alias{nextTabButton}
\alias{prevTabButton}
\title{Navigation sidebar}
\usage{
navSidebar(id, class = "navSidebar", .prev = TRUE, .next = TRUE, ...)

quitButton(id)

saveButton(id)

nextTabButton(id)

prevTabButton(id)
}
\arguments{
\item{id}{shiny id}

\item{class}{css class style. Default is set to "navSidebar":
.navSidebar {
  width: 100%;
  text-align: center;
  padding: 0;
}}

\item{.prev}{logical. Do you want a "Previous" actionButton? (prevTabButton)}

\item{.next}{logical. Do you want a "Next" actionButton? (nextTabButton)}

\item{...}{UI tags set at the bottom of the navigation side bar}
}
\description{
UI part for shiny navigation sidebar module
}
\section{Functions}{
\itemize{
\item \code{quitButton}: Quit button.

\item \code{saveButton}: Save button.

\item \code{nextTabButton}: Next button.

\item \code{prevTabButton}: Prev button.
}}

