% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eal_0_functions.R
\name{navSidebarUI}
\alias{navSidebarUI}
\alias{quitButton}
\alias{saveButton}
\alias{nextTabButton}
\alias{prevTabButton}
\title{navSidebarUI}
\usage{
navSidebarUI(id, class = "navSidebar", .prev = TRUE, .next = TRUE)

quitButton(id)

saveButton(id)

nextTabButton(id)

prevTabButton(id)
}
\arguments{
\item{id}{shiny id}

\item{class}{css class style. Default is set to "navSidebar":
.navSidebar {
  width: 100%;
  text-align: center;
  padding: 0;
}}

\item{.prev}{logical. Do you want a "Previous" actionButton? (prevTabButton)}

\item{.next}{logical. Do you want a "Next" actionButton? (nextTabButton)}

\item{...}{UI tags set at the bottom of the navigation side bar}
}
\description{
UI part for shiny navigation sidebar module
}
\section{Functions}{
\itemize{
\item \code{quitButton}: 

\item \code{saveButton}: 

\item \code{nextTabButton}: 

\item \code{prevTabButton}: 
}}

