% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL_functions.R
\name{nextTab}
\alias{nextTab}
\title{nextTab}
\usage{
nextTab(input, output, session, globals, previous)
}
\arguments{
\item{globals}{global variable containing application settings}

\item{previous}{character. Name of the recently quit page.}
}
\description{
server side of the 'nextTabButton' UI. Allows to
increment a specific variable contained in globals by 1.
}
