% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-doc-documentation_functions.R
\name{nt.titles}
\alias{nt.titles}
\title{nt.titles}
\usage{
nt.titles(vec, action_target)
}
\arguments{
\item{vec}{"mother vector" containing elements}

\item{action_target}{duo of named elements and 'renderUI' action to perform}
}
\description{
Apply a precise action on elements from _vec_ named after _vec_'s names
}
