% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL_functions.R
\name{onSave}
\alias{onSave}
\title{onSave}
\usage{
onSave(input, output, session, toSave, path, filename)
}
\arguments{
\item{toSave}{R objects to save}

\item{path}{directory location where the file will be saved}

\item{filename}{name of the file to save}
}
\description{
server side of the 'saveButton' UI
}
