% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-isContentTruthy.R
\name{optional}
\alias{optional}
\title{optional}
\usage{
optional(x, type = NULL)
}
\arguments{
\item{x}{value provided for the argument}

\item{type}{typed value (e.g. character()) in case option is invalid.}
}
\description{
Fill optional arguments of a function if the given argument comes to be not
truthy (e.g. by getting it after a script execution and passing it to a
function).
}
