% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL_personnel.R
\name{PersonnelMod}
\alias{PersonnelMod}
\title{PersonnelMod}
\usage{
PersonnelMod(input, output, session, globals, rv, rmv_id, site_id, ref)
}
\description{
server part for 'Personnel' module
}
