% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eal_7_personnel.R
\name{PersonnelModUI}
\alias{PersonnelModUI}
\alias{PersonnelMod}
\title{PersonnelModUI}
\usage{
PersonnelModUI(id, div.id, site.id, rmv.id, role = NULL, saved = NULL)

PersonnelMod(
  input,
  output,
  session,
  main.env,
  rv,
  rmv.id,
  site.id,
  ref,
  role = NULL,
  saved = NULL
)
}
\description{
module to document EML Personnel
}
\section{Functions}{
\itemize{
\item \code{PersonnelMod}: 
}}

