% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL_functions.R
\name{printReactiveValues}
\alias{printReactiveValues}
\title{printReactiveValues}
\usage{
printReactiveValues(values)
}
\arguments{
\item{values}{reactiveValues. The reactiveValues object to return}
}
\description{
clean way to display reactiveValues. This only works
with one-level-deep reactiveValues
}
