% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_readDataTable.R
\name{readDataTable}
\alias{readDataTable}
\title{Read Data Tables}
\usage{
readDataTable(file, data.table = FALSE, ...)
}
\arguments{
\item{file}{path to the data table file. For Excel files,
supports "http://", "https://", and "ftp://" URLS.}

\item{data.table}{(from \code{fread()}) TRUE returns a data.table.
FALSE returns a data.frame. The default for this argument can
be changed with options(datatable.fread.datatable=FALSE).}

\item{...}{additional arguments to read.table.}
}
\value{
a data.frame
}
\description{
Disclaimer: this function is still maturing. There is no guarantee about its
performance.
Guess the type of the data file (e.g. .xls* or not) and loads it accordingly.
}
