% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_readPlaintext.R
\name{readPlainText}
\alias{readPlainText}
\title{readFilesText}
\usage{
readPlainText(files, prefix = NULL, sep = "/", ...)
}
\arguments{
\item{files}{files basename located in the same directory or,
if prefix = NULL, list of full filenames to read}

\item{prefix}{common file prefix for all file names
specified in 'files'. By default, sep = "/"}
}
\description{
readFilesText
}
