% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_additional_HTML.R
\name{remove_shiny_inputs}
\alias{remove_shiny_inputs}
\title{Clear backstage shiny observers}
\usage{
remove_shiny_inputs(id, .input)
}
\arguments{
\item{id}{character. An ID string that corresponds with the ID used to call
the module's UI function.*}

\item{.input}{internal. Shiny server \code{input} variable passed to servers.}
}
\description{
Clear server-side of a shiny module
}
\details{
Freely teached from a community soluce on
\href{https://appsilon.com/how-to-safely-remove-a-dynamic-shiny-module/}{appsilon}.
}
