% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-about-about_functions.R
\name{renderBibliography}
\alias{renderBibliography}
\title{renderBibliography}
\usage{
renderBibliography(bib)
}
\arguments{
\item{bib}{file path to bibliography}
}
\description{
shiny-formatted render* function. Allow the user to print a .bib bibliography content. References are NOT numbered
according to possible calls from the app.
}
