% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_rightSideBar.R
\name{rightSidebarSettings}
\alias{rightSidebarSettings}
\title{settings in rightSideBar}
\usage{
rightSidebarSettings(id, wip)
}
\description{
UI part of the settings module. Allow the user to change several
settings in the app. This is designed to be wrapped in {shinydashboardPlus}.
}
