% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_run.R
\name{runMetashark}
\alias{runMetashark}
\title{Run MetaShARK}
\usage{
runMetashark(...)
}
\arguments{
\item{...}{options to pass to the application, ignored if missing or mistyped.
\describe{
\item{wip}{logical. Shows WIP parts of the app. (default to FALSE)}
\item{dev}{logical. Add development elements in the GUI.
(default to FALSE)}
\item{reactlog}{logical. Use reactlog? (default to TRUE)}
\item{use_profvis}{logical. Profile the app with {profvis}?
(default to FALSE)}
\item{use_test}{logical. Lanch the app with its testing environment?
(default to FALSE). Shall not be used by end-users.}
}}
}
\description{
Main function for launching the MetaShARK application.
}
\details{
MetaShARK (METAdata SHiny Automated Resource & Knowledge) is a web app
which is designed to help its user as much as possible for filling ecological
metadata. It uses the EML standard (cf. NCEAS work) to allow a full and
precise description of input datasets.
}
\examples{

## Not run:

library(MetaShARK)

# run this to launch MetaShARK
runMetashark()

## End (Not run)

}
\author{
Elie Arnaud \href{mailto:elie.arnaud@mnhn.fr}{elie.arnaud@mnhn.fr}
}
