% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL_selectDP.R
\name{selectDP}
\alias{selectDP}
\title{Data Package selection}
\usage{
selectDP(input, output, session, savevar, globals)
}
\description{
UI part for the Data Package selection. Allow the user to choose between
creating a new data package or loading an existing one.
}
