% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL_templateDP.R
\name{templateDPUI}
\alias{templateDPUI}
\alias{templateDP}
\title{Data Package Template filling}
\usage{
templateDPUI(id, title, dev)

templateDP(input, output, session, savevar, globals)
}
\description{
UI part of the templateDP module. Fill in the attributes of the data package

server part of the templateDP module.
}
\section{Functions}{
\itemize{
\item \code{templateDP}: 
}}

