% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-upload-upload.R
\name{upload}
\alias{upload}
\title{upload}
\usage{
upload(input, output, session, dev, globals)
}
\arguments{
\item{input}{shiny module input}

\item{output}{shiny module output}

\item{session}{shiny module session}

\item{dataone.formats}{list of the dataone supported formats}
}
\description{
server part of the upload module
}
