% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-upload-functions.R
\name{uploadDP}
\alias{uploadDP}
\title{uploadDP}
\usage{
uploadDP(mn, cn, eml, data, scripts = c(), formats)
}
\arguments{
\item{mn}{Member Node URL.}

\item{cn}{Contributing Node URL.}

\item{eml}{Metadata file path (XML validating NCEAS' EML Schema)}

\item{data}{Data files path}

\item{scripts}{Scripts files path}

\item{formats}{List of DataONE CN supported [formats](https://cn.dataone.org/cn/v2/formats)}
}
\description{
automation script to upload data packages
produced with {dataone} and {datapack}. This function is 
freely inspired by [this tutorial script](http://training.arcticdata.io/materials/arctic-data-center-training/programming-metadata-and-data-publishing.html#publish-data-to-the-arctic-data-center-test-site).
}
