% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload-module.R
\name{uploadUI}
\alias{uploadUI}
\alias{upload}
\title{uploadUI}
\usage{
<<<<<<< HEAD
uploadUI(id)
=======
uploadUI(id, main.env)
>>>>>>> 21780e3c7e17505ab12284e63b960fbb7e749dc8

upload(id, main.env)
}
\arguments{
\item{id}{shiny module id}

\item{main.env}{inner global environment}
}
\description{
UI part for the upload module. Used to build and drop
data packages to a chosen metacat.
}
\section{Functions}{
\itemize{
\item \code{upload}: 
}}

