% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_additional_HTML.R
\name{wipRow}
\alias{wipRow}
\title{wipRow}
\usage{
wipRow(...)
}
\arguments{
\item{...}{content to style}
}
\value{
styled content as WIP
}
\description{
styles an inputRow to give it WIP appearance
}
