% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-writeText.R
\name{write.text}
\alias{write.text}
\title{write.text}
\usage{
write.text(x, file = ".", collapse = "\\n")
}
\arguments{
\item{x}{A character vector or list. A list will be unlisted. If length(x) > 1
then every item will be collapsed into a single vector.}

\item{file}{A file name (1-length character vector).}
}
\description{
Applies 'writeLines' in a friendly way. This function has been
quickly designed to allow the writing of plain text files.
}
