% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL_Attributes.R
\name{AttributesUI}
\alias{AttributesUI}
\alias{Attributes}
\title{Data Package Template filling}
\usage{
AttributesUI(id, title, dev)

Attributes(input, output, session, savevar, globals, NSB)
}
\description{
UI part of the Attributes module. Fill in the attributes of the data package

server part of the Attributes module.
}
\section{Functions}{
\itemize{
\item \code{Attributes}: 
}}

