% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_additional_HTML.R
\name{checkFeedback}
\alias{checkFeedback}
\title{checkFeedback

relies on {shinyFeedback} to automate feedback on one input.}
\usage{
checkFeedback(
  input,
  id,
  condition = NULL,
  silent = FALSE,
  type = c("danger", "warning"),
  text = NULL
)
}
\arguments{
\item{id}{character. An ID string that corresponds with the ID used to call
the module's UI function.}

\item{condition}{logical. Determines if the feedback is positive or not.}

\item{silent}{logical. If TRUE, the only feedback occurs when \code{condition} is
met. If FALSE, also displays feedbacks for other cases.}

\item{type}{character. Either "danger" or "warning", sets the type of
feedback in case \code{condition} is not met.}

\item{text}{character. What message to display in case of unmet condition.}
}
\description{
checkFeedback

relies on {shinyFeedback} to automate feedback on one input.
}
