% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-collapsible.R
\name{collapsibleUI}
\alias{collapsibleUI}
\alias{collapsible}
\title{collapsibleUI}
\usage{
collapsibleUI(id, label, hidden = TRUE, ...)

collapsible(input, output, session)
}
\arguments{
\item{label}{A label appearing on the clickable link.}

\item{...}{any UI element}
}
\description{
A module to get a div collapsed by clicking on a link.
}
\section{Functions}{
\itemize{
\item \code{collapsible}: 
}}

