% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-collapsible.R
\name{collapsibleUI}
\alias{collapsibleUI}
\alias{collapsible}
\title{collapsibleUI}
\usage{
collapsibleUI(id, label, hidden = TRUE, ..., class = NULL)

collapsible(input, output, session)
}
\arguments{
\item{label}{character. A label appearing on the clickable link.}

\item{...}{shiny UI elements. Any UI element displayed as core content.}

\item{class}{character. CSS class to apply to ... .}
}
\description{
A shiny module to get a div collapsed by clicking on a link.
}
\section{Functions}{
\itemize{
\item \code{collapsible}: 
}}

