% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-collapsible.R
\name{collapsibleUI}
\alias{collapsibleUI}
\alias{collapsible}
\title{collapsibleUI}
\usage{
collapsibleUI(id, label, .hidden = TRUE, ..., class = NULL)

collapsible(id)
}
\arguments{
\item{id}{character. An ID string that corresponds with the ID used to call
the module's UI function.}

\item{label}{character. A label appearing on the clickable link.}

\item{.hidden}{logical. A flag to make the UI display as collapsed or not.}

\item{...}{shiny UI elements. Any UI element displayed as core content.}

\item{class}{character. CSS class to apply to ... .}
}
\description{
A shiny module to get a div collapsed by clicking on a link.
}
\section{Functions}{
\itemize{
\item \code{collapsible}: Server part for collapsible widget module.
}}

