% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eal_0.R
\name{EMLAL}
\alias{EMLAL}
\title{EMLAL}
\usage{
EMLAL(input, output, session, savevar, globals)
}
\description{
server part of the EMLAL module. Allow the user to use a front-end shiny interface to the EML Assembly Line package, from
Environmental Data Initiative.
}
