% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-followPath.R
\name{followPath}
\alias{followPath}
\title{followPath}
\usage{
followPath(tree, path, sep = "/")
}
\arguments{
\item{tree:}{explored hierarchy list thanks to @path}

\item{path:}{vector of characters matching some of @tree names and}

\item{sep:}{separators between @path elements (aka @tree names)}
}
\description{
Takes a hierarchy list (tree), a path written in a vector pasted
with sep = sep, and returns the targetted node
}
