% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eal_5_GeoCovInput.R
\name{GeoCovInputUI}
\alias{GeoCovInputUI}
\alias{GeoCovInput}
\title{Geographic coverage modular Input}
\usage{
GeoCovInputUI(id, site.id, rmv.id, default = NULL)

GeoCovInput(input, output, session, rv, rmv.id, site.id, ref)
}
\description{
UI part for the Geographic Coverage modular input, used
in Geographic Coverage module to input a custom entry in the dataset.
}
\section{Functions}{
\itemize{
\item \code{GeoCovInput}: 
}}

