% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL_GeoCovInput.R
\name{GeoCovInputUI}
\alias{GeoCovInputUI}
\alias{GeoCovInput}
\title{Geographic coverage modular Input}
\usage{
GeoCovInputUI(id, site_id, rmv_id)

GeoCovInput(input, output, session, rv, rmv_id, site_id, ref)
}
\description{
UI part for the Geographic Coverage modular input, used
in Geographic Coverage module to input a custom entry in the dataset.
}
\section{Functions}{
\itemize{
\item \code{GeoCovInput}: 
}}

