% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-%grep%.R
\name{\%grep\%}
\alias{\%grep\%}
\title{%grep%}
\usage{
x \%grep\% table
}
\arguments{
\item{x}{vector or NULL: the values to be matched as a pattern.}

\item{table}{vector or NULL: the values to be matched against. Long vectors
are not supported.
@param ... additional argument valid for `grepl`.}
}
\description{
This function is a copycat from `%in%` to use grepl instead of exact match. 
If `x` is found in a part of any element of `table`, returns TRUE.
}
