% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-module_functions.R
\name{initReactive}
\alias{initReactive}
\alias{saveReactive}
\title{initReactive}
\usage{
initReactive(sublist = NULL, savevar = NULL, glob)

saveReactive(savevar, rv = NULL, .write = FALSE)
}
\arguments{
\item{sublist}{either NULL, "emlal", "metafin" to precise which sublist to initialize (NULL initializes the whole variable)}
}
\value{
`savevar` modified.
}
\description{
EMLAL module specific function. Initialize `savevar` variable.

save the `savevar` variable at wanted location
}
\section{Functions}{
\itemize{
\item \code{saveReactive}: 
}}

