% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-fill-fill_functions.R
\name{initReactive}
\alias{initReactive}
\title{initReactive}
\usage{
initReactive(sublist = NULL, savevar = NULL)
}
\arguments{
\item{savevar}{NULL if sublist is NULL, existing savevar else.}

\item{sublist:}{either NULL, "emlal", "metafin". Define if the
whole variable must be set (NULL) or if a sublist shall be reset}
}
\description{
Create a savevar reactiveValues (specifically used in
MetaShARK) or a sublist of savevar. The sublists of savevar are
created accordingly to the "little r strategy" from ThinkR
(https://rtask.thinkr.fr/fr/tag/strategie-petit-r/)
}
