% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-module_functions.R
\name{initReactive}
\alias{initReactive}
\alias{saveReactive}
\title{initReactive}
\usage{
initReactive(sub.list = NULL, save.variable = NULL, main.env)

saveReactive(save.variable, rv = NULL, main.env = NULL)
}
\arguments{
\item{sub.list}{either NULL, "emlal", "metafin" to precise which sub.list to initialize (NULL initializes the whole variable)}
}
\value{
`save.variable` modified.
}
\description{
EMLAL module specific function. Initialize `save.variable` variable.

save the `save.variable` variable at wanted location
}
\section{Functions}{
\itemize{
\item \code{saveReactive}: 
}}

