% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_isContentTruthy.R
\name{isContentTruthy}
\alias{isContentTruthy}
\title{Check for data structures validity}
\usage{
isContentTruthy(x)
}
\arguments{
\item{x}{argument to check fo truthiness}
}
\value{
A logical indicating whether or not the variable is Truthy.
}
\description{
Check if \code{x} is truthy (as shiny::isTruthy) or not, but also
checks for its potential content. This function uses \code{unlist()} and \code{all()} to
check every bit of the variable.
}
