% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_listReactiveValues.R
\name{listReactiveValues}
\alias{listReactiveValues}
\title{List Reactive Values}
\usage{
listReactiveValues(rv, lv = 0, name = "root")
}
\arguments{
\item{rv}{reactiveValues. Target to turn into list.}

\item{lv}{internal. Verbose purposes.}

\item{name}{internal. Current root node name}
}
\description{
Allows to turn a `reactiveValues`` object into a non-reactive list. Uses
recursive method.
}
