% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-reactiveTrigger.R
\name{makeReactiveTrigger}
\alias{makeReactiveTrigger}
\title{reactiveTrigger}
\usage{
makeReactiveTrigger()
}
\description{
create a 'reactiveTrigger' object (NOT a proper R class) with two methods:
1. depend() : must be written in a code chunk to execute on triggering
2. trigger() : when executed, trigger the object (and all the "depending" code chunks)
This function is freely reused from Dean Attali's work.
}
