% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-markdownInput.R
\name{markdownInputUI}
\alias{markdownInputUI}
\alias{markdownInput}
\title{markdownInputUI}
\usage{
markdownInputUI(id, label = "Text", icon = TRUE, preview = FALSE, value = "")

markdownInput(id, preview = FALSE)
}
\arguments{
\item{id}{character. The input slot that will be used to access the value.}

\item{label}{character. Display label for the control, or NULL for no label.}

\item{icon}{character. An optional `icon()` to appear on the button.}

\item{preview}{logical. Shall a preview panel appear? (right-sided, 50\% width)}

\item{value}{character. Initial value of the text input (default to "")}
}
\value{
A HTML `reactive` containing the formatted input. Do call it as a reactive.
}
\description{
A shiny input module designed to allow the user to write markdown and render it as
a HTML fix.
}
\section{Functions}{
\itemize{
\item \code{markdownInput}: Server side of the module.
}}

\examples{

ui <- fluidPage(
  markdownInputUI("md", label = "Write markdown here", preview = TRUE)
)

server <- function(input, output, session) {
  callModule(markdownInput, "md", preview = TRUE)
}

shinyApp(ui, server)
}
\references{
Ace technologies \url{https://ace.c9.io/#nav=about}
}
