% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-multiFilesInput.R
\name{multiFilesInputUI}
\alias{multiFilesInputUI}
\alias{multiFilesInput}
\title{multiFilesInputUI}
\usage{
multiFilesInputUI(id, helpText = NULL, server = TRUE)

multiFilesInput(input, output, session, server = TRUE)
}
\arguments{
\item{id}{shiny module id}

\item{helpText}{a character vector to give the user an explanation. Will be inserted into a tags$p() call.}

\item{input}{shiny module input}

\item{output}{shiny module output}

\item{session}{shiny module session}
}
\description{
Shiny module to allow the user to chose multiple files,
and remove all or part of his selection.
}
\section{Functions}{
\itemize{
\item \code{multiFilesInputUI}: 

\item \code{multiFilesInput}: 
}}

