% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL_functions.R
\name{navSidebar}
\alias{navSidebar}
\title{navSideBar}
\usage{
navSidebar(id, class = "navSidebar", .prev = TRUE, .next = TRUE, ...)
}
\arguments{
\item{.prev}{logical. Shall 'previous' button appear?}

\item{.next}{logical. Shall 'next' button appear?}
}
\value{
A shiny tagList.
}
\description{
container thought to embed Shiny tags allowing navigation
in a shiny app.
}
