% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eal_0_functions.R
\name{navSidebar}
\alias{navSidebar}
\alias{onQuit}
\alias{onSave}
\alias{nextTab}
\alias{prevTab}
\title{navSidebar}
\usage{
navSidebar(id, globals, savevar)

onQuit(input, output, session, globals, savevar, NSB)

onSave(input, output, session, savevar, NSB)

nextTab(input, output, session, globals, NSB)

prevTab(input, output, session, globals, NSB)
}
\arguments{
\item{id}{shiny server id}

\item{globals}{MetaShARK globals variable}

\item{savevar}{MetaShARK savevar variable}
}
\description{
server part for shiny navigation sidebar module (see [navSidebar()]).
The functions are very specific and thus are not exported.
}
\section{Functions}{
\itemize{
\item \code{onQuit}: 

\item \code{onSave}: 

\item \code{nextTab}: 

\item \code{prevTab}: 
}}

