% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL_functions.R
\name{onQuit}
\alias{onQuit}
\title{onQuit}
\usage{
onQuit(input, output, session, globals, toSave, path, filename)
}
\arguments{
\item{globals}{global variable containing application settings}

\item{toSave}{R objects to save}

\item{path}{directory location where the file will be saved}

\item{filename}{name of the file to save}
}
\description{
server part of the quitButton UI.
}
