% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL_functions.R
\name{onQuit}
\alias{onQuit}
\alias{onSave}
\alias{nextTab}
\alias{prevTab}
\title{Navigation server}
\usage{
onQuit(input, output, session, globals, toSave, path, filename)

onSave(input, output, session, toSave, path, filename)

nextTab(input, output, session, globals, previous)

prevTab(input, output, session, globals, previous)
}
\description{
server part for shiny navigation sidebar module (see [navSidebar()]). 
The functions are very specific and thus are not exported.
}
\section{Functions}{
\itemize{
\item \code{onSave}: 

\item \code{nextTab}: 

\item \code{prevTab}: 
}}

