% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about_module_functions.R
\name{renderBibliography}
\alias{renderBibliography}
\title{renderBibliography}
\usage{
renderBibliography(bib)
}
\arguments{
\item{bib}{character. Path to bibliography file (.bib format).}
}
\description{
shiny-formatted render* function. Allow the user to print a .bib bibliography content. References are NOT numbered
according to possible calls from the app.
}
