% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMetashark.R
\name{runMetashark}
\alias{runMetashark}
\title{Run MetaShARK}
\usage{
runMetashark(...)
}
\arguments{
\item{...}{options to pass to the application, ignored if missing or mistyped.
\describe{
  \item{wip}{logical. Shows WIP parts of the app. (default to FALSE)}
  \item{dev}{logical. Add development elements in the GUI. (default to FALSE)}
  \item{reactlog}{logical. Use reactlog? (default to TRUE)}
}}
}
\description{
Main function for launching the MetaShARK application.
}
\details{
MetaShARK (METAdata SHiny Automated Resource & Knowledge) is a web app
which is designed to help its user as much as possible for filling ecological
metadata. It uses the EML standard (cf. NCEAS work) to allow a full and
precise description of input datasets.
}
\examples{
# run this to launch MetaShARK
runMetashark()

}
\author{
Elie Arnaud <elie.arnaud@mnhn.fr>
}
