% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL_selectDP.R
\name{selectDP}
\alias{selectDP}
\title{selectDP}
\usage{
selectDP(input, output, session, savevar, globals)
}
\arguments{
\item{savevar}{global reactiveValue containing the saved information
entered by the user.}

\item{globals}{global list containing fixed setting values for the
app.}
}
\description{
server part of the EML AL module (step 1: load an existing
data package OR create a new one)
}
