% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL_SelectDP.R
\name{SelectDP}
\alias{SelectDP}
\title{Data Package selection}
\usage{
SelectDP(input, output, session, savevar, globals)
}
\description{
UI part for the Data Package selection. Allow the user to choose between
creating a new data package or loading an existing one.
}
