% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL_selectDP.R
\name{selectDPUI}
\alias{selectDPUI}
\title{selectDPUI}
\usage{
selectDPUI(id, title, width = 12, dev = FALSE)
}
\arguments{
\item{title}{page title}

\item{width}{same width as in shiny::column}

\item{dev}{logical. Shall the dev items appear?}
}
\description{
UI part of the EML AL module (step 1: load an existing
data package OR create a new one)
}
