% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL_SelectDP.R
\name{SelectDPUI}
\alias{SelectDPUI}
\title{Data Package selection}
\usage{
SelectDPUI(id, title, width = 12, dev = FALSE, server)
}
\description{
UI part for the Data Package selection. Allow the user to choose between
creating a new data package or loading an existing one.
}
