% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eal_1_SelectDP.R
\name{SelectDPUI}
\alias{SelectDPUI}
\alias{loadFromDP}
\title{Data Package selection}
\usage{
SelectDPUI(id, width = 12, dev = FALSE)

loadFromDP(save.variable, file)
}
\description{
UI part for the Data Package selection. Allow the user to choose between
creating a new data package or loading an existing one.
}
\section{Functions}{
\itemize{
\item \code{loadFromDP}: Sets up a DP from a pre-written EML file.
}}

