% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-emlal-EMLAL_templateDP.R
\name{templateDPUI}
\alias{templateDPUI}
\title{templateDPUI}
\usage{
templateDPUI(id, title, dev)
}
\arguments{
\item{title}{page title}

\item{dev}{logical. Shall the dev items appear?}
}
\description{
UI part of the EML AL module (step 3: fill in
the data attributes)
}
