% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-upload-upload.R
\name{uploadUI}
\alias{uploadUI}
\title{uploadUI}
\usage{
uploadUI(id, dev)
}
\arguments{
\item{id}{shiny module id}
}
\description{
UI part for the upload module. Used to build and drop
data packages to a chosen metacat.
}
