% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-URL_Input.R
\name{URL_Input_UI}
\alias{URL_Input_UI}
\alias{URL_Input}
\title{URL_Input_UI}
\usage{
URL_Input_UI(id, label = "URL", width = "100\%")

URL_Input(input, output, session)
}
\value{
(output of `callModule`) If input is a valid URL (regex-tested + curl-tested), returns input.
Else, returns NA.
}
\description{
URL Shiny input module.
}
\section{Functions}{
\itemize{
\item \code{URL_Input}: 
}}

