% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_withRedStar.R
\name{withRedStar}
\alias{withRedStar}
\title{withRedStar}
\usage{
withRedStar(text)
}
\arguments{
\item{text}{the HTLM text to put before the red star}
}
\value{
an html element
}
\description{
Add a red star at the end of the text

Adds a red star at the end of the text
(for example for indicating mandatory fields).
}
\examples{
withRedStar("Enter your name here")
}
